refactoring: scala3 pc - extract completions into a separate package
