+++ mtags/src/main/scala-3/scala/meta/internal/mtags/MtagsEnrichments.scala
@@ -18,6 +18,7 @@ import dotty.tools.dotc.interactive.Interactive
 import dotty.tools.dotc.interactive.InteractiveDriver
 import dotty.tools.dotc.util.SourcePosition
 import dotty.tools.dotc.util.Spans
+import org.eclipse.lsp4j.MarkupContent
 import org.eclipse.{lsp4j as l}
 
 object MtagsEnrichments extends CommonMtagsEnrichments:
