+++ mtags/src/main/scala-3/scala/meta/internal/pc/AutoImportsProvider.scala
@@ -8,6 +8,7 @@ import scala.jdk.CollectionConverters.*
 
 import scala.meta.internal.mtags.MtagsEnrichments.*
 import scala.meta.internal.pc.AutoImports.*
+import scala.meta.internal.pc.completions.CompletionPos
 import scala.meta.pc.*
 
 import dotty.tools.dotc.ast.tpd.*
